Add ability not to follow selected item in the table
It is not always useful, consider csysdig, by default when you did not
press anything it stick to the first item by sort order, and only if you
select something (Up/Down) then it will stick to this item in the table.

And I also made the Home button to cancel this selection.